store/requirements.txt
django
djangorestframework
psycopg2
requests

